backend/requirements.txt
flask==3.0.2
pymongo==4.6.2
python-dotenv==1.0.1